Currycarbon.CLI.RunCalibrate

Documentation

data CalibrateOptions

A data type to represent the options to the CLI module function runCalibrate

Constructors

CalibrateOptions

Fields

_calibrateUncalC14 :: [UncalC14]

Uncalibrated dates that should be calibrated

_calibrateUncalC14File :: [FilePath]

List of files with uncalibrated dates to be calibrated

_calibrateCalCurveFile :: Maybe FilePath

Path to a .14c file

_calibrateCalibrationMethod :: CalibrationMethod

Calibration algorithm that should be used

_calibrateAllowOutside :: Bool

Allow calibration to run outside of the range of the calibration curve

_calibrateDontInterpolateCalCurve :: Bool

Don't interpolate the calibration curve

_calibrateQuiet :: Bool

Suppress the printing of calibration results to the command line

_calibrateDensityFile :: Maybe FilePath

Path to an output file (see CLI documentation)

_calibrateHDRFile :: Maybe FilePath

Path to an output file

_calibrateCalCurveSegmentFile :: Maybe FilePath

Path to an output file

_calibrateCalCurveMatrixFile :: Maybe FilePath

Path to an output file

runCalibrate :: CalibrateOptions -> IO ()

Interface function to trigger calibration from the command line