Currycarbon.Calibration.Calibration

Calibration functions

getRelevantCalCurveSegment :: UncalC14 -> CalCurveBP -> CalCurveBP

Take an uncalibrated date and a raw calibration curve and return the relevant segment of the calibration curve

prepareCalCurveSegment :: Bool -> CalCurveBP -> CalCurveBCAD

Modify a calibration curve (segment) with multiple optional steps, including interpolation and transforming dates to BC/AD format

makeCalCurveMatrix :: UncalPDF -> CalCurveBCAD -> CalCurveMatrix

Construct a matrix representation of a calibration curve for a given date

uncalToPDF :: UncalC14 -> UncalPDF

Transform an uncalibrated date to an uncalibrated probability density table

calibrateDate

Arguments

:: CalibrateDatesConf
Configuration options to consider

-> CalCurveBP
A calibration curve

-> UncalC14
An uncalibrated radiocarbon date

-> Either CurrycarbonException CalPDF
The function returns either an exception if the calibration failed for some reason, or a CalPDF

Calibrates a date with the provided calibration curve

calibrateDates

Arguments

:: CalibrateDatesConf
Configuration options to consider

-> CalCurveBP
A calibration curve

-> [UncalC14]
A list of uncalibrated radiocarbon dates

-> [Either CurrycarbonException CalPDF]
The function returns a list for each input date, with either an exception if the calibration failed for some reason, or a CalPDF

Calibrates a list of dates with the provided calibration curve

refineCalDates :: [CalPDF] -> [Maybe CalC14]

Transforms the raw, calibrated probability density table to a meaningful representation of a calibrated radiocarbon date

refineCalDate :: CalPDF -> Maybe CalC14

data CalibrateDatesConf

A data type to cover the configuration options of the calibrateDates function

Constructors

CalibrateDatesConf

Fields

_calConfMethod :: CalibrationMethod

The calibration algorithm that should be used

_calConfAllowOutside :: Bool

Allow calibration to run outside of the range of the calibration curve

_calConfInterpolateCalCurve :: Bool

Interpolate the calibration curve before calibration. This is a simple linear interpolation only to increase the output resolution for earlier time periods, where the typical calibration curves are less dense by default. With the interpolation, the output will be a per-year density. The mechanism is inspired by the implementation in the Bchron R package

defaultCalConf :: CalibrateDatesConf

A default configuration that should yield almost identical calibration results to the Bchron R package