Currycarbon.Calibration.MatrixMult

Documentation

calibrateDateMatrixMult :: Bool -> Bool -> CalCurve -> UncalC14 -> Either CurrycarbonException CalPDF

makeCalCurveMatrix :: UncalPDF -> CalCurve -> CalCurveMatrix

Construct a matrix representation of a calibration curve for a given date

uncalToPDF :: UncalC14 -> UncalPDF

Transform an uncalibrated date to an uncalibrated probability density table