Currycarbon.Calibration.Utils

Documentation

normalizeCalPDF :: CalPDF -> CalPDF

Rescale a CalPDF so that the sum of the densities is approx. 1.0

dnorm :: Float -> Float -> Float -> Float

get the density of a normal distribution at a point x

dt :: Double -> Float -> Float

get the density of student's-t distribution at a point x

isOutsideRangeOfCalCurve :: CalCurveBP -> UncalC14 -> Bool

getRelevantCalCurveSegment :: UncalC14 -> CalCurveBP -> CalCurveBP

Take an uncalibrated date and a raw calibration curve and return the relevant segment of the calibration curve

prepareCalCurveSegment :: Bool -> CalCurveBP -> CalCurveBCAD

Modify a calibration curve (segment) with multiple optional steps, including interpolation and transforming dates to BC/AD format

makeBCADCalCurve :: CalCurveBP -> CalCurveBCAD

vectorBPToBCAD :: Vector YearBP -> Vector YearBCAD

interpolateCalCurve :: CalCurveBP -> CalCurveBP

trimLowDensityEdgesCalPDF :: CalPDF -> CalPDF