Currycarbon.SumCalibration

Documentation

evalCalExpr :: CalibrateDatesConf -> CalCurveBP -> CalExpr -> Either CurrycarbonException CalPDF

Evaluate a dating expression by calibrating the individual dates and forming the respective sums and products of post-calibration density distributions

eitherCombinePDFs :: (Float -> Float -> Float) -> Float -> Either CurrycarbonException CalPDF -> Either CurrycarbonException CalPDF -> Either CurrycarbonException CalPDF

addPDFs :: CalPDF -> CalPDF -> CalPDF

Add two probabilty densities

multiplyPDFs :: CalPDF -> CalPDF -> CalPDF

Multiply two probabilty densities

combinePDFs :: (Float -> Float -> Float) -> Float -> CalPDF -> CalPDF -> CalPDF