Currycarbon.Types

Data types

This module defines the relevant data types for handling radiocarbon dates

data CalibrationMethod

Different calibration algorithms implemented in currycarbon. Currently two distinct implementations are available, although both of them are similar Intercept calibration algorithms. Maybe more algorithms will be added in the future

Constructors

MatrixMultiplication
A matrix multiplication method generally following this blog post by Martin Hinz. This method is slower and the underlying code more verbose than Bchron, but it has some advantages regarding didactics and the inspection of intermediate data products for debugging. Using this method is thus generally not advisable, except for specific applications, where a more technical insight into C14 calibration is needed

Bchron
A fast and reliable calibration algorithm very similar to the implementation in the R package Bchron by Andrew Parnell. This algorithm can be run with a simple normal distribution (NormalDist) or Student's t-distribution (StudentTDist), which is recommended

Fields

distribution :: CalibrationDistribution

data CalibrationDistribution

Statistical distributions to be used with the CalibrationMethod Bchron

Constructors

NormalDist
Normal distribution

StudentTDist
Student's t-distribution.

Fields

ndf :: Double

number of degrees of freedom

type YearBP = Word

A type to represent years BP. All numbers are positive and describe the distance in years to 1950AD: 3000 = 3000BP = 1050BC

type YearBCAD = Int

A type to represent years BC or AD. Negative values describe years BC, positive values years AD: -5000 = 5000BC and 1300 = 1300AD

type YearRange = Word

A type to represent a range of years

data UncalC14

A data type to represent an uncalibrated radiocarbon date

Constructors

UncalC14

Fields

_uncalC14Id :: String

Sample identifier, e.g. a lab number

_uncalC14UnCal :: YearBP

C14 age in years BP

_uncalC14Sigma :: YearRange

C14 standard deviation (one sigma in years)

data UncalPDF

A data type to represent a year-wise probability density for uncalibrated dates Although technically not correct, we still call this a probability density function (PDF)

Constructors

UncalPDF

Fields

_uncalPDFid :: String

Sample identifier, e.g. a lab number

_uncalPDFUnCals :: Vector YearBP

Years BP

_uncalPDFDens :: Vector Float

Probability densities

data CalCurveBP

A data type to represent a calibration curve with YearBP

Constructors

CalCurveBP

Fields

_calCurveBPCals :: Vector YearBP

Years calBP

_calCurveBPUnCals :: Vector YearBP

Years BP

_calCurveBPSigmas :: Vector YearRange

Standard deviation (one sigma in years)

data CalCurveBCAD

A second data type to represent a calibration curve, here now with YearBCAD

Constructors

CalCurveBCAD

Fields

_calCurveBCADCals :: Vector YearBCAD

Years calBCAD

_calCurveBCADUnCals :: Vector YearBCAD

Years BCAD

_calCurveBCADSigmas :: Vector YearRange

Standard deviation (one sigma in years)

data CalCurveMatrix

A data type to represent a calibration curve in a wide matrix form

Constructors

CalCurveMatrix

Fields

_calCurveMatrixUnCals :: Vector YearBCAD

Row names of the calibration curve matrix: Years BCAD

_calCurveMatrixCals :: Vector YearBCAD

Column names of the calibration curve matrix: Years calBCAD

_calCurveMatrixDens :: Vector (Vector Float)

Matrix (as a list of columns) with the probability densities

data CalPDF

A data type to represent a year-wise probability density for calibrated dates. Although technically not correct, we still call this a probability density function (PDF)

Constructors

CalPDF

Fields

_calPDFid :: String

Sample identifier, e.g. a lab number

_calPDFCals :: Vector YearBCAD

Years calBCAD

_calPDFDens :: Vector Float

Probability densities for each year in _calPDFCals

data CalExpr

A data type to represent an expression for sum- or product calibration

Constructors

UnCalDate UncalC14

CalDate CalPDF

SumCal CalExpr CalExpr

ProductCal CalExpr CalExpr

data CalC14

A data type to represent a human readable summary of a calibrated radiocarbon date

Constructors

CalC14

Fields

_calC14id :: String

Identifier, e.g. a lab number

_calC14RangeSummary :: CalRangeSummary

Summary of the range of the calibrated date

_calC14HDROneSigma :: [HDR]

One-sigma high density regions

_calC14HDRTwoSigma :: [HDR]

Two-sigma high density regions

data CalRangeSummary

A data type to store a simple range summary of the calibrated date, including the median age

Constructors

CalRangeSummary

Fields

_calRangeStartTwoSigma :: YearBCAD

Start of the two-sigma age range

_calRangeStartOneSigma :: YearBCAD

Start of the one-sigma age range

_calRangeMedian :: YearBCAD

Median age

_calRangeStopOneSigma :: YearBCAD

End of the one-sigma age range

_calRangeStopTwoSigma :: YearBCAD

End of the two-sigma age range

data HDR

A data type to represent a high density region of a probability distribution. A high density region is here defined as an age range, within which the respective cummulative probability (e.g. of an calibrated radiocarbon date density curve) is above a certain threshold

Constructors

HDR

Fields

_hdrstart :: YearBCAD

Start of the high density region in years calBCAD

_hdrstop :: YearBCAD

End of the high density region in years calBCAD