Currycarbon.Utils

Documentation

data CurrycarbonException

Different exceptions for currycarbon

Constructors

CurrycarbonCLIParsingException String
An exception to describe an issue in the currycarbon CLI input parsing

CurrycarbonCalibrationRangeException String
An exection to describe the case that a date is not in the range of the supplied calibration curve

renderCurrycarbonException :: CurrycarbonException -> String