Currycarbon.Parsers

Parsing, rendering and writing functions

This module contains a number of functions to manage data input and output plumbing for different datatypes

renderCalDatesPretty :: [(UncalC14, CalC14, CalPDF)] -> String

renderCalDatePretty :: (UncalC14, CalC14, CalPDF) -> String

readCalibrationMethod :: String -> Either String CalibrationMethod

parseCalibrationMethod :: Parser CalibrationMethod

writeCalC14s :: FilePath -> [CalC14] -> IO ()

renderCalC14ForFile :: CalC14 -> String

renderCalC14s :: [CalC14] -> String

renderCalC14 :: CalC14 -> String

renderHDRsForFile :: [HDR] -> [(String, String)]

renderHDRForFile :: HDR -> (String, String)

renderHDRs :: [HDR] -> String

renderHDR :: HDR -> String

writeCalCurveMatrix :: FilePath -> CalCurveMatrix -> IO ()

renderCalCurveMatrix :: CalCurveMatrix -> String

writeCalPDFs :: FilePath -> [CalPDF] -> IO ()

renderBuilderCalPDF :: CalPDF -> Builder

renderCLIPlotCalPDF :: Int -> Int -> CalPDF -> String

renderUncalC14 :: UncalC14 -> String

readUncalC14FromFile :: FilePath -> IO [UncalC14]

readUncalC14 :: String -> Either String [UncalC14]

parseUncalC14 :: Parser UncalC14

writeCalCurve :: FilePath -> CalCurve -> IO ()

renderCalCurve :: CalCurve -> String

readCalCurveFromFile :: FilePath -> IO CalCurve

readCalCurve :: String -> CalCurve

parseCalCurve :: Parser [(Int, Int, Int)]

parseCalCurveLine :: Parser (Int, Int, Int)

comments :: Parser String