Currycarbon.Types

Data types

This module defines the relevant data types for handling radiocarbon dates

data CalibrationMethod

Different calibration algorithms implemented in currycarbon. Currently two distinct implementations are available, although both of them are similar Intercept calibration algorithms. Maybe more algorithms will be added in the future

Constructors

MatrixMultiplication
A matrix multiplication method generally following this blog post by Martin Hinz. This method is slower and the underlying code more verbose than Bchron, but it has some advantages regarding didactics and the inspection of intermediate data products for debugging. Using this method is thus generally not advisable, except for specific applications, where a more technical insight into C14 calibration is needed

Bchron
A fast and reliable calibration algorithm very similar to the implementation in the R package Bchron by Andrew Parnell. This algorithm can be run with a simple normal distribution (NormalDist) or Student's t-distribution (StudentTDist), which is recommended

Fields

distribution :: CalibrationDistribution

data CalibrationDistribution

Statistical distributions to be used with the CalibrationMethod Bchron

Constructors

NormalDist
Normal distribution

StudentTDist
Student's t-distribution.

Fields

ndf :: Double

number of degrees of freedom

data UncalC14

A data type to represent an uncalibrated radiocarbon date

Constructors

UncalC14

Fields

_uncalC14Id :: String

Sample identifier, e.g. a lab number

_uncalC14BP :: Int

C14 age in years BP

_uncalC14Sigma :: Int

C14 standard deviation (one sigma in years)

data UncalPDF

A data type to represent a year-wise probability density for uncalibrated dates Although technically not correct, we still call this a probability density function (PDF)

Constructors

UncalPDF

Fields

_uncalPDFid :: String

Sample identifier, e.g. a lab number

_uncalPDFBPs :: Vector Int

Years BP or BC

_uncalPDFDens :: Vector Float

Probability densities

data CalCurve

A data type to represent a calibration curve

Constructors

CalCurve

Fields

_calCurveCals :: Vector Int

Years calBP or calBC

_calCurveBPs :: Vector Int

Years BP or BC

_calCurveSigmas :: Vector Int

Standard deviation (one sigma in years)

data CalCurveMatrix

A data type to represent a calibration curve in a wide matrix form

Constructors

CalCurveMatrix

Fields

_calCurveMatrixBPs :: Vector Int

Row names of the calibration curve matrix: Years BP or BC

_calCurveMatrixCalBPs :: Vector Int

Column names of the calibration curve matrix: Years calBP or calBC

_calCurveMatrixDens :: Vector (Vector Float)

Matrix (as a list of columns) with the probability densities

data CalPDF

A data type to represent a year-wise probability density for calibrated dates. Although technically not correct, we still call this a probability density function (PDF)

Constructors

CalPDF

Fields

_calPDFid :: String

Sample identifier, e.g. a lab number

_calPDFBPs :: Vector Int

Calibrated years calBP or calBC/AD (depending on the context)

_calPDFDens :: Vector Float

Probability densities for each year in _calPDFBPs

data CalC14

A data type to represent a human readable summary of a calibrated radiocarbon date

Constructors

CalC14

Fields

_calC14id :: String

Identifier, e.g. a lab number

_calC14HDROneSigma :: [HDR]

One-sigma high density regions

_calC14HDRTwoSigma :: [HDR]

Two-sigma high density regions

data HDR

A data type to represent a high density region of a probability distribution. A high density region is here defined as an age range, within which the respective cummulative probability (e.g. of an calibrated radiocarbon date density curve) is above a certain threshold

Constructors

HDR

Fields

_hdrstart :: Int

Start of the high density region

_hdrstop :: Int

End of the high density region