Currycarbon.Types

Data types

This module defines the relevant data types for handling radiocarbon dates

data UncalC14

A data type to represent an uncalibrated radiocarbon date

Constructors

UncalC14

Fields

_uncalC14id :: String

Identifier, e.g. a lab number

_uncalC14age :: Int

Uncalibrated age BP in years

_uncalC14sd :: Int

One-sigma standard deviation of the uncalibrated age in years

data UncalPDF

A data type to represent a year-wise probability density for uncalibrated dates Although technically not correct, we still call this a probability density function

Constructors

UncalPDF

Fields

_uncalPDFid :: String

Identifier, e.g. a lab number

_uncalPDFagedens :: [(Int, Float)]

List that stores years BP and the respective probability densities alongside each other

getBPsUncal :: UncalPDF -> [Int]

Getter function to access the year BP variable in an uncalibrated age probability density list

getProbsUncal :: UncalPDF -> [Float]

Getter function to access the density variable in an uncalibrated age probability density list

newtype CalCurve

A data type to represent a calibration curve

Constructors

CalCurve [(Int, Int, Int)]
List that stores the sequence of year BP, year calBP and the one-sigma standard deviation (on the year BP scale) that makes up the calibration curve

getBPs :: CalCurve -> [Int]

Getter function to access the year BP variable in a calibration curve

getCals :: CalCurve -> [Int]

Getter function to access the year calBP variable in a calibration curve

getCalSigmas :: CalCurve -> [Int]

Getter function to access the sigma variable in a calibration curve

data CalCurveMatrix

A data type to represent a calibration curve in a wide matrix form

Constructors

CalCurveMatrix

Fields

_calCurveMatrixbp :: [Int]

Row names of the calibration curve matrix: Uncalibrated age BP in years

_calCurveMatrixcalbp :: [Int]

Column names of the calibration curve matrix: ^ Calibrated age BP in years

_calCurveMatrixmatrix :: [[Float]]

Matrix (as a list of columns) with the probaility densities

data CalPDF

A data type to represent a year-wise probability density for calibrated ages

Constructors

CalPDF

Fields

_calPDFid :: String

Identifier, e.g. a lab number

_calPDFagedens :: [(Int, Float)]

List that stores years calBP and the respective probability densities alongside each other

getBPsCal :: CalPDF -> [Int]

Getter function to access the year BP variable in a calibrated age probability density list

getProbsCal :: CalPDF -> [Float]

Getter function to access the density variable in a calibrated age probability density list

data CalC14

A data type to represent a human readable summary of a calibrated radiocarbon date

Constructors

CalC14

Fields

_calC14id :: String

Identifier, e.g. a lab number

_calC14hdronesigma :: [HDR]

One-sigma high density regions

_calC14hdrtwosigma :: [HDR]

Two-sigma high density regions

data HDR

A data type to represent a high density region of a probability distribution. A high density region is here defined as an age range, within which the respective cummulative probability (e.g. of an calibrated radiocarbon date density curve) is above a certain threshold

Constructors

HDR

Fields

_hdrstart :: Int

Start of the high density region

_hdrstop :: Int

End of the high density region